Try a really random node.
Previously the function failed against the same node again and again as
the logic to test the node and put it at the end of the list is still
not here. This is just a temporary solution, a fixed list with
non-responding nodes moved at the end is probably better in the long
run.