consistency-test: write more often to a subset of keys.
Instead of a single key space size, after this commit there are two: the
working set and the set of keys.

For example if the working set is set to 100 and the key space to 1
million, the program will write half the time to the first 100 keys, and
half time time evently across all the keys.

This way it is easy to spot consistency issues without too much delay
but at the same time it is possible to work with a greater set of keys
in order to simulate big reshardings.